+++ example/src/component/button/button.module.css
@@ -1,0 +1,24 @@
+.button {
+  display: inline-block;
+  padding: 11px 20px;
+  outline: 0;
+  border: 1px solid #ccc;
+  border-radius: 3px;
+  background-color: #fff;
+  box-shadow: 0 1px 1px 0 rgba(100, 100, 100, 0.2);
+  color: #555;
+  font-size: 16px;
+  text-align: center;
+  cursor: pointer;
+  white-space: nowrap;
+}
+
+.button:active {
+  background-color: #fafafa;
+}
+
+.button:disabled {
+  background-color: #eee;
+  color: #aaa;
+  cursor: not-allowed;
+}
