+++ example/src/index.css
@@ -1,9 +1,22 @@
 @import url('https://fonts.googleapis.com/css2?family=Patrick+Hand+SC&display=swap');
 
+:root {
+  --basic-font-size: 17px;
+}
+
 * {
   box-sizing: border-box;
   margin: 0;
   padding: 0;
+}
+
+html,
+body {
+  height: 100%;
+}
+
+#root {
+  height: 100%;
 }
 
 body {
@@ -13,6 +26,7 @@ body {
   font-family: -apple-system, system-ui, BlinkMacSystemFont, 'Segoe UI', Roboto,
     'Helvetica Neue', Arial, sans-serif;
   color: #444;
+  font-size: var(--basic-font-size);
   line-height: 1.5;
 }
 
@@ -34,89 +48,27 @@ ol {
   margin-bottom: 80px;
 }
 
-.links {
-  position: absolute;
-  top: 3px;
-  right: 30px;
-}
-
-.links a {
-  color: #5d657a;
-  font-size: 15px;
-  text-decoration: none;
-}
-
-.links a + a {
-  margin-left: 20px;
-}
-
-footer {
-  border-top: 1px solid #eee;
-}
-
-footer .footer-content {
-  width: 1100px;
-  max-width: 100%;
-  margin: 0 auto;
-  padding: 50px 30px 70px;
-}
-
-footer a {
-  color: #999;
-  text-decoration: none;
-}
-
-/*.container {*/
-/*  width: 1100px;*/
-/*  max-width: 100%;*/
-/*  margin: 0 auto;*/
-/*  padding: 40px 30px 120px;*/
-/*}*/
-
 h2 {
-  margin-bottom: 30px;
-  padding: 50px 0 20px;
-  border-bottom: 1px solid #eee;
-  color: #555;
-  font-size: 40px;
+  margin-bottom: 20px;
+  font-size: 30px;
 }
 
 h3 {
   margin: 70px 0 20px;
-  color: #555;
-  font-size: 28px;
+  font-size: 22px;
 }
 
 p {
-  margin-bottom: 30px;
-  color: #666;
-  font-size: 18px;
+  font-size: var(--basic-font-size);
 }
 
-.example-button {
-  display: inline-block;
-  min-width: 240px;
-  padding: 12px 20px;
-  outline: 0;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  background-color: #fff;
-  box-shadow: 0 1px 1px 0 rgba(100, 100, 100, 0.2);
-  color: #555;
-  font-size: 16px;
-  text-align: center;
-  cursor: pointer;
-  white-space: nowrap;
+p a {
+  color: darkgoldenrod;
+  text-decoration: underline;
 }
 
-.example-button:active {
-  background-color: #fafafa;
-}
-
-.example-button:disabled {
-  background-color: #eee;
-  color: #aaa;
-  cursor: not-allowed;
+section {
+  margin-bottom: 50px;
 }
 
 .example-button + .example-button {
