+++ example/src/type/utils.ts
@@ -1,0 +1,1 @@
+export type AtLeast<T, K extends keyof T> = Partial<T> & Pick<T, K>
